Rename README.md to ABOUT MEn                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      About ME
 About Me

Name: Sir Taevon Garner

Hometown: Atlanta, GA

IT Interest: What interests me most about IT currently are the massive amount of opportunities currently available for those who take their time with getting acquainted what IT is now and understanding what IT has to offer going into the future. 
As someone who has a history in physical security and a passion for the internet and IT now being able to advance myself and my lineage within the IT realm is a blessing I cannot take for granted.

Career Plans/Goals:
I aspire to not only obtain both certificates but to also fully grasp the concept of Networking, Security protocols, Python, Linux and eventually coding in-depth. After a few goals of I have include obtaining my Gvt secret top security clearance and obtaining a career/contract working with Major companies like Lockheed Martin, DOD, Delta, Boeing, and or any of the many similar major organization that holds the same importance; as a cyber security expert who is not only knowledgable but also continuously growing within the cybersec field and IT altogether. 
After completing the cybersecurity course with City of Refuge, I aim to be placed with a organization where I learn further and gain the needed experience to work with a company or organization courtesy of the job placement faculty, somewhere I can grow with fro. novice to expert and potentially Ethical Hacker or CYSA expert within the Gvt.

Proud Moment of 2025:
I am very proud of the opportunity and experience I have in front of myself through the City of Refuge. The unique opportunity it presents along with what the organization represents and offers in a City like Atlanta; giving those who are aware but not as financially equipped with an opportunity of a lifetime. If possible, I would love to become an ambassador or in some way give back to those that believed and invested in me.

-God is Great Always-